requirements.txt
# ── Web app ───────────────────────────────────────────────
streamlit>=1.35.0

# ── Git / repo handling ────────────────────────────────────
gitpython>=3.1.41

# ── AI (Groq API — free, fast) ─────────────────────────────
groq>=0.9.0

# ── AI (Ollama — local fallback) ───────────────────────────
ollama>=0.2.0

# ── Visualisation ─────────────────────────────────────────
plotly>=5.20.0

# ── Data ──────────────────────────────────────────────────
pandas>=2.0.0

# ── Utilities ─────────────────────────────────────────────
chardet>=5.2.0

# ── Generative AI ───────────────
google-generativeai>=0.3.0